Shuffling things around
* move configuration of producer for cloud to cloud config
* get rid of Cloud::Instrumenter and rely on subscribe of configured
instrumenter
* move event to cloud namespace from cloud instrumenter

This should make it a little more easy to inject producer and test it on
its own.

The downside is that an instrumenter is required for events. I imagine
most people use active support notifications through rails so that
shouldn't matter. If it does, we can make our own down the road.